chore: update all i18n.t to hooks
