+++ src/constants/HealthRecordingConstants.ts
@@ -1,5 +1,3 @@
-import i18n from '../internationalization/i18n.config';
-
 export enum TimeFrame {
   WEEK = 'week',
   TWO_WEEKS = '2week',
@@ -8,33 +6,6 @@ export enum TimeFrame {
   YEAR = 'year',
   ALL_TIME = 'allTime'
 }
-
-export const timeFrameOption = [
-  {
-    label: i18n.t('analytics.week'),
-    value: 'week'
-  },
-  {
-    label: i18n.t('analytics.2week'),
-    value: '2week'
-  },
-  {
-    label: i18n.t('analytics.1month'),
-    value: 'month'
-  },
-  {
-    label: i18n.t('analytics.3month'),
-    value: '3months'
-  },
-  {
-    label: i18n.t('analytics.year'),
-    value: 'year'
-  },
-  {
-    label: i18n.t('analytics.allTime'),
-    value: 'allTime'
-  }
-];
 
 export const chartConfig = {
   backgroundColor: '#e26a00',
